Fix sitemap and extract sorting+label lookup
