+++ proxy/bin/buildAssets.ts
@@ -1,16 +1,8 @@
 import * as fs from 'fs'
 import { getMenuItemsFromRepoItems } from '../src/utils'
+import { getFileLabel, sortFilesByStructure } from './fileTree'
 import { includeActions } from './libReferences/actions/actions'
 import { includeFormulas } from './libReferences/formulas/formulas'
-
-interface MenuItemStructure {
-  path: string
-  label: string
-}
-
-export type MenuItemsStructure = Array<
-  MenuItemStructure & { children?: Array<MenuItemStructure> }
->
 
 // Setup
 if (fs.existsSync('./dist/docs')) {
@@ -35,69 +27,17 @@ const actionReferencePath = './dist/docs/references/actions/index.md'
 const actionReferenceContent = fs.readFileSync(actionReferencePath, 'utf-8')
 const newActionContent = await includeActions(actionReferenceContent)
 fs.writeFileSync(actionReferencePath, newActionContent, 'utf-8')
-// Structure that holds the order of all menu items/content
-const structure = JSON.parse(
-  fs.readFileSync('../docs/index.json', 'utf-8'),
-) as MenuItemsStructure
 
 // Read all files and sort them according to the structure from docs/index.json
-const files = (
-  fs.readdirSync('./dist/docs', { recursive: true }) as string[]
-).sort((a, b) => {
-  const aParts = a.split('/')
-  const bParts = b.split('/')
-  const getIndexString = (parts: string[]) => {
-    const folderIndex = structure.findIndex((item) => item.path === parts[0])
-    if (folderIndex === -1) {
-      // eslint-disable-next-line no-console
-      console.warn(`Folder ${parts[0]} not found in structure`)
-      throw new Error(`Folder ${parts[0]} not found in index.json`)
-    }
-    const folderKey = `${folderIndex}`.padStart(3, '0')
-
-    if (parts.length > 1 && parts[1] !== 'index.md') {
-      const childIndex = structure[folderIndex].children?.findIndex(
-        (item) => item.path === parts[1],
-      )
-      if (childIndex === -1) {
-        // eslint-disable-next-line no-console
-        console.warn(`Child path ${parts[1]} not found in folder ${parts[0]}`)
-        throw new Error(
-          `Child path ${parts[1]} not found in folder ${parts[0]}`,
-        )
-      }
-      const childKey = `${childIndex}`.padStart(3, '0')
-      return `${folderKey}.${childKey}`
-    }
-    return `${folderKey}`
-  }
-  return getIndexString(aParts).localeCompare(getIndexString(bParts))
-})
-const getLabel = (path: string) => {
-  const parts = path.split('/')
-  const folder = structure.find((item) => item.path === parts[0])
-  if (!folder) {
-    // eslint-disable-next-line no-console
-    console.warn(`Folder ${parts[0]} not found in structure`)
-    throw new Error(`Folder ${parts[0]} not found in index.json`)
-  }
-  if (parts.length > 1 && parts[1] !== 'index.md') {
-    const child = folder.children?.find((item) => item.path === parts[1])
-    if (!child) {
-      // eslint-disable-next-line no-console
-      console.warn(`Child path ${parts[1]} not found in folder ${parts[0]}`)
-      throw new Error(`Child path ${parts[1]} not found in folder ${parts[0]}`)
-    }
-    return child.label
-  }
-  return folder.label
-}
+const files = sortFilesByStructure(
+  fs.readdirSync('./dist/docs', { recursive: true }) as string[],
+)
 // Create menu items structure
 const menuItems = getMenuItemsFromRepoItems({
   items: files.map((file) => ({
     path: `docs/${file}`,
     type: 'tree',
-    label: getLabel(file),
+    label: getFileLabel(file),
   })),
   parentPath: 'docs',
   owner: 'nordcraftengine',
